www: Extract PendingBuildRequestsTable component
